requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Pillow==10.1.0
rembg==2.0.50
PyPDF2==3.0.1
pdf2docx==0.5.8